requirements_mapas_interativos.txt
streamlit-image-coordinates
Pillow
